--- src/api/parser/src/routes/parser.js
@@ -1,9 +1,0 @@
-const { Router, logger } = require('@senecacdot/satellite');
-
-const router = Router();
-
-router.get('/', (req, res) => {
-  res.status(200).send('hi');
-});
-
-module.exports = router;
