+++ src/api/parser/src/server.js
@@ -1,4 +1,3 @@
-const { logger, fetch } = require('@senecacdot/satellite');
 const service = require('.');
 
 const port = parseInt(process.env.PARSER_PORT || 10000, 10);
